feat: set is_sent -field to additional_data as False on log creation
KK-1413.

Whenever a new LogEntry instance is created (and is preparing to be
saved to db), it should have the `additional_data` field populated with
`is_sent` boolean (set to False). This is a requirement of
https://github.com/City-of-Helsinki/structured-log-transfer. When the
structured log transfer tool finishes handling the data, it will then
mark the LogEntry as sent (by updating
`additional_data["is_sent"]=True`).